refactor: :truck: Rename NotificationOutlet file
index.tsx -> NotificationOutlet.tsx
