+++ src/renderer/components/NotificationsOutlet/NotificationsOutlet.tsx
