+++ src/renderer/components/Outlet/Outlet.tsx
@@ -5,7 +5,7 @@ import CustomRouter from '../CustomRouter/CustomRouter';
 import * as Styled from './style';
 import { selectCurrentView } from '../../../shared/redux/slices/currentViewSlice';
 import { FULL_SCREEN_VIEWS } from '../../constants/RouterComponents';
-import NotificationsOutlet from '../NotificationsOutlet';
+import NotificationsOutlet from '../NotificationsOutlet/NotificationsOutlet';
 
 const Outlet = () => {
   const { view } = useSelector(selectCurrentView);
